initial notebooks and data
refactored_atm_rad_schmit-newturbulence-altitude notebook uses the turbulence class (in the turbulence_class notebook) as well as the altitude to calculate the turbulent heating rate. this is the most up to date of the notebooks. schmit data is in monthly datasets over 1994-2012; the schmit_compiled_temp_data_final is the compilation of this data into one excel; the schmit_compiled_gas_data is the final compilation of the gas data into one excel.